requirements.txt
absl_py==1.2.0
catboost==1.2
cvxpy==1.3.2
ddl==0.0.2
ite==1.4.2
joblib==1.2.0
kneed==0.7.0
lightgbm==3.3.2
matplotlib==3.4.2
metrics==0.3.3
numpy==1.22.4
pandas==1.5.3
POT==0.9.0
pykdtree==1.3.7.post0
scikit_learn==1.2.2
scipy==1.11.3
setuptools==67.7.2
setuptools==65.5.1
tensorflow==2.11.0
tensorflow_datasets==4.9.3
tensorflow_hub==0.12.0
torch==1.9.0
torch==2.0.1
torchvision==0.9.1
util==0.0.1
xgboost==1.7.6
